core: Refactor SharedObject
